+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Expect active development and potentially significant breaking changes in the `0
 - Fix: make sure maybeDeepFreeze is called on results returned from setVariables and refetch [PR #1362](https://github.com/apollographql/apollo-client/pull/1362)
 - Fix: use setTimeout to throw uncaught errors in observer.next and observer.error[PR #1367](https://github.com/apollographql/apollo-client/pull/1367)
 - Remove returnPartialData option [PR #1370](https://github.com/apollographql/apollo-client/pull/1370)
+- Implement fetchPolicy to replace noFetch and forceFetch [PR #1371](https://github.com/apollographql/apollo-client/pull/1371)
 
 
 ### 0.10.1